Slightly rework auth, only require login for specific components
