+++ frontend/components/auth/PrivyLoginButton.tsx
@@ -1,0 +1,57 @@
+import { useLogin } from "@privy-io/react-auth";
+import { usePrivy } from "@privy-io/react-auth";
+import React from "react";
+import { useDispatch } from "react-redux";
+
+import { ButtonProps } from "@/components/ui/button";
+import { Button } from "@/components/ui/button";
+import { AppDispatch, setIsLoggedIn, setWalletAddress } from "@/lib/redux";
+import { saveUserAsync } from "@/lib/redux/slices/userSlice/thunks";
+
+const PrivyLoginButton = (props: ButtonProps) => {
+  const dispatch: AppDispatch = useDispatch();
+  const { ready, authenticated } = usePrivy();
+
+  //This component must remain mounted wherever you use it for the callback to fire correctly
+  const { login } = useLogin({
+    onComplete: (user, isNewUser, wasAlreadyAuthenticated) => {
+      const walletAddress = user?.wallet?.address;
+      if (!walletAddress) {
+        console.log("No wallet address found");
+        return;
+      }
+      dispatch(setWalletAddress(walletAddress));
+
+      if (wasAlreadyAuthenticated) {
+        console.log("User was already authenticated");
+        dispatch(setIsLoggedIn(true));
+      } else if (isNewUser) {
+        console.log("New user");
+        dispatch(saveUserAsync({ walletAddress }));
+        dispatch(setIsLoggedIn(true));
+      } else if (user) {
+        console.log("User authenticated");
+        dispatch(setIsLoggedIn(true));
+      }
+    },
+    onError: (error) => {
+      console.log("onError callback triggered", error);
+    },
+  });
+
+  const handleLogin = async () => {
+    try {
+      login();
+    } catch (error) {
+      console.log("Error calling login function:", error);
+    }
+  };
+
+  return ready && !authenticated ? (
+    <Button onClick={handleLogin} {...props}>
+      Log In
+    </Button>
+  ) : null;
+};
+
+export default PrivyLoginButton;
